# Remove StringBuilder marshaling from GetNameInfoW