--- a/src/System.Net.NameResolution/src/System/Net/NameResolutionPal.Windows.cs
+++ b/src/System.Net.NameResolution/src/System/Net/NameResolutionPal.Windows.cs
@@ -243,42 +243,32 @@ namespace System.Net
             return SocketError.Success;
         }
 
-        public static string TryGetNameInfo(IPAddress addr, out SocketError errorCode, out int nativeErrorCode)
+        public static unsafe string TryGetNameInfo(IPAddress addr, out SocketError errorCode, out int nativeErrorCode)
         {
-            //
-            // Use SocketException here to show operation not supported
-            // if, by some nefarious means, this method is called on an
-            // unsupported platform.
-            //
-            SocketAddress address = (new IPEndPoint(addr, 0)).Serialize();
-            StringBuilder hostname = new StringBuilder(1025); // NI_MAXHOST
-
-            int flags = (int)Interop.Winsock.NameInfoFlags.NI_NAMEREQD;
-
-            nativeErrorCode = 0;
-
-            byte[] addressBuffer = new byte[address.Size];
+            SocketAddress address = new IPEndPoint(addr, 0).Serialize();
+            Span<byte> addressBuffer = address.Size <= 64 ? stackalloc byte[64] : new byte[address.Size];
             for (int i = 0; i < address.Size; i++)
             {
                 addressBuffer[i] = address[i];
             }
 
-            errorCode =
-                Interop.Winsock.GetNameInfoW(
-                    addressBuffer,
+            const int NI_MAXHOST = 1025;
+            char* hostname = stackalloc char[NI_MAXHOST];
+
+            nativeErrorCode = 0;
+            fixed (byte* addressBufferPtr = addressBuffer)
+            {
+                errorCode = Interop.Winsock.GetNameInfoW(
+                    addressBufferPtr,
                     address.Size,
                     hostname,
-                    hostname.Capacity,
+                    NI_MAXHOST,
                     null, // We don't want a service name
                     0, // so no need for buffer or length
-                    flags);
-
-            if (errorCode != SocketError.Success)
-            {
-                return null;
+                    (int)Interop.Winsock.NameInfoFlags.NI_NAMEREQD);
             }
 
-            return hostname.ToString();
+            return errorCode == SocketError.Success ? new string(hostname) : null;
         }
 
         public static unsafe string GetHostName()