cc: @JeremyKuhne, @davidsh, @jkotas 

```C#
[Benchmark] public IPHostEntry GetHostEntry() => Dns.GetHostEntry("34.206.253.53");
```

Before:
```
       Method |     Mean |    Error |   StdDev | Allocated |
------------- |---------:|---------:|---------:|----------:|
 GetHostEntry | 717.8 us | 14.20 us | 41.21 us |   2.73 KB |
```

After:
```
       Method |     Mean |    Error |   StdDev | Allocated |
------------- |---------:|---------:|---------:|----------:|
 GetHostEntry | 708.1 us | 14.02 us | 34.12 us |     624 B |
```